feat: implemented authentication and added tests to validate implementation

+++ backend/src/main/java/dev/levelupschool/backend/service/ArticleService.java
@@ -1,13 +1,15 @@
 package dev.levelupschool.backend.service;
 
+import dev.levelupschool.backend.auth.AuthenticationUtils;
 import dev.levelupschool.backend.exception.ModelNotFoundException;
 import dev.levelupschool.backend.model.Article;
+import dev.levelupschool.backend.model.User;
 import dev.levelupschool.backend.repository.ArticleRepository;
 import dev.levelupschool.backend.repository.UserRepository;
-import dev.levelupschool.backend.request.CreateArticleRequest;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.nio.file.AccessDeniedException;
 import java.util.List;
 
 @Service
@@ -28,30 +30,41 @@ public class ArticleService {
             .orElseThrow(() -> new ModelNotFoundException(Article.class, id));
     }
 
-    public Article createArticle(CreateArticleRequest request) {
-        Long userId = request.getUserId();
-        String title = request.getTitle();
-        String content = request.getContent();
+    public Article createArticle(Article article) {
+        User loggedInUser = AuthenticationUtils.getLoggedInUser(userRepository);
 
-        return articleRepository.save(
-            new Article(
-                title,
-                content,
-                userRepository.findById(userId).orElseThrow(() -> new ModelNotFoundException(Article.class, userId))
-            )
-        );
+        article.setAuthor(loggedInUser);
+
+        return articleRepository.save(article);
     }
 
-    public Article updateArticle(Article newArticle, Long id) {
-        return articleRepository.findById(id)
-            .map(article -> {
-                article.setTitle(newArticle.getTitle());
-                article.setContent(newArticle.getContent());
-                return articleRepository.save(article);
-            }).orElseThrow(() -> new ModelNotFoundException(Article.class, id));
+    public Article updateArticle(Article newArticle, Long id) throws AccessDeniedException {
+        var loggedInUser = AuthenticationUtils.getLoggedInUser(userRepository);
+
+        var article = articleRepository.findById(id)
+            .orElseThrow(() -> new ModelNotFoundException(Article.class, id));
+
+        if (!article.getAuthor().getId().equals(loggedInUser.getId())) {
+            throw new AccessDeniedException("You cannot edit an article not created by you!");
+        }
+
+        article.setTitle(newArticle.getTitle());
+        article.setContent(newArticle.getContent());
+
+        return articleRepository.save(article);
     }
 
-    public void deleteArticle(Long id) {
+    public void deleteArticle(Long id) throws AccessDeniedException {
+        var loggedInUser = AuthenticationUtils.getLoggedInUser(userRepository);
+
+        var article = articleRepository
+            .findById(id)
+            .orElseThrow(() -> new ModelNotFoundException(Article.class, id));
+
+        if (!loggedInUser.getId().equals(article.getAuthor().getId())) {
+            throw new AccessDeniedException("You cannot delete an article not created by you!");
+        }
+
         articleRepository.deleteById(id);
     }
 }
